Code organization tweaks.
