--- assets/css/dashboard/new-pa-message-page/where-card.scss
@@ -1,20 +1,0 @@
-.where-card {
-  margin-top: 24px;
-  padding: 18px 24px;
-  background-color: $cool-gray-20;
-  color: white;
-
-  .title {
-    margin-bottom: 16px;
-    font-weight: 500;
-    font-size: 24px;
-  }
-
-  .add-stations-zones-button {
-    background-color: $text-link-primary;
-    color: $text-invert;
-    border: 1px dashed $text-invert;
-    width: fit-content;
-    vertical-align: middle;
-  }
-}
